Move alert-related mappers into more consistent packages
Fix HS-1783
