+++ alert/src/main/java/org/opennms/horizon/alertservice/mapper/AlertMapper.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.alertservice.service;
+package org.opennms.horizon.alertservice.mapper;
 
 import java.util.Date;
